requirements.txt
gradio
langchain
langchain_groq
Pillow
python-dotenv